Analysis of CNN's tweets using LDA
This is a indivisual project undertaken by me because of my interest in ML and NLP.

The projects involves collection of CNN's tweets (more than 3200)  over the year and analyzing them using LDA(latent dirichlet allocation) as weel as some other features.

The results of the analysis is mentioned in the report.